feat: Refactor delete-user flow, add helpers module & improve middleware usage
