+++ account-service/docs/swagger.json
@@ -268,9 +268,6 @@
                 "full_name": {
                     "type": "string"
                 },
-                "has_card": {
-                    "type": "boolean"
-                },
                 "password": {
                     "type": "string"
                 },
