+++ account-service/routers/account-router.go
@@ -1,6 +1,7 @@
 package routers
 
 import (
+	"account-service/middleware"
 	"account-service/service"
 
 	"github.com/gin-gonic/gin"
@@ -9,7 +10,17 @@ import (
 func AccountRouter(r *gin.RouterGroup, accountService *service.AccountService) {
 	r.GET("/accounts", accountService.GetAccounts)
 	r.GET("/fetch-acc/:acc_num", accountService.GetAccoutById)
+
+	// Adding new User
 	r.POST("/register", accountService.AddAccount)
+
+	// Update has_card field
 	r.PATCH("/update_acc/:acc_num", accountService.UpdateHasCard)
-	r.DELETE("/delete-user/:email", accountService.DeleteUser)
+
+	// Delete User
+	r.DELETE("/delete-user/:email",
+		middleware.FetchUserUUID(),
+		middleware.DeleteAuthUser(),
+		accountService.DeleteUser,
+	)
 }
