+++ account-service/routers/swagger.go
@@ -19,11 +19,11 @@ func InitializeSwagger(r *gin.Engine) {
 	docs.SwaggerInfo.Title = "Bank-It Account Service Documentation API"
 	docs.SwaggerInfo.Description = "This is an api Testing and Documentation for Bank-It services."
 	docs.SwaggerInfo.Version = "1.0"
-	docs.SwaggerInfo.Host = "localhost:8081"
+	docs.SwaggerInfo.Host = "localhost:8082"
 	docs.SwaggerInfo.BasePath = "/api/account-service"
 	docs.SwaggerInfo.Schemes = []string{"http", "https"}
 
 	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
 
-	logger.Info("Swagger running on http://localhost:8081/swagger/index.html")
+	logger.Info("Swagger running on http://localhost:8082/swagger/index.html")
 }
